+++ djangocms_blog/__init__.py
@@ -1,6 +1,8 @@
 # -*- coding: utf-8 -*-
+from __future__ import absolute_import, print_function, unicode_literals
+
 __author__ = 'Iacopo Spalletti'
 __email__ = 'i.spalletti@nephila.it'
-__version__ = '0.8.0.b8'
+__version__ = '0.9a1'
 
 default_app_config = 'djangocms_blog.apps.BlogAppConfig'