+++ HISTORY.rst
@@ -1,10 +1,62 @@
 .. :changelog:
 
+=======
 History
--------
+=======
 
-0.8.0 (unreleased)
-++++++++++++++++++
+******************
+0.9.0 (unreleased)
+******************
+
+* Dropped support for Django<1.8, django CMS<3.2.
+* Added liveblog application.
+* Refactored plugin filters: by default only data for current site are now shown.
+* Added global and per site posts count to BlogCategory.
+* Added option to hide empty categories from menu.
+* Added standalone documentation at https://djangocms-blog.readthedocs.io.
+* Enabled cached version of BlogLatestEntriesPlugin.
+* Added plugins templateset.
+* Improved category admin to avoid circular relationships.
+
+******************
+0.8.6 (unreleased)
+******************
+
+* Set the correct language during indexing
+
+******************
+0.8.5 (2016-06-26)
+******************
+
+* Fixed issues with ThumbnailOption migration under mysql.
+
+******************
+0.8.4 (2016-06-22)
+******************
+
+* Fixed issues with cmsplugin-filer 1.1.
+
+******************
+0.8.3 (2016-06-21)
+******************
+
+* Stricter filer dependency versioning.
+
+******************
+0.8.2 (2016-06-12)
+******************
+
+* Aldryn-only release. No code changes
+
+******************
+0.8.1 (2016-06-11)
+******************
+
+* Aldryn-only release. No code changes
+
+******************
+0.8.0 (2016-06-05)
+******************
 
 * Added django-knocker integration
 * Changed the default value of date_published to null
@@ -16,9 +68,11 @@ History
 * Changed slug size to 255 chars
 * Fixed pagination setting in list views
 * Added API to set default sites if user has permission only for a subset of sites
+* Added Aldryn integration
 
+******************
 0.7.0 (2016-03-19)
-++++++++++++++++++
+******************
 
 * Make categories non required
 * Fix tests with parler>=1.6
@@ -33,29 +87,33 @@ History
 * Mitigate issue when apphook config can't be retrieved
 * Mitigate issue when wizard double registration is triggered
 
+******************
 0.6.3 (2015-12-22)
-++++++++++++++++++
+******************
 
 * Add BLOG_ADMIN_POST_FIELDSET_FILTER to filter admin fieldsets
 * Ensure correct creation of full URL for canonical urls
 * Move constants to settings
 * Fix error when no config is found
 
+******************
 0.6.2 (2015-11-16)
-++++++++++++++++++
+******************
 
 * Add app_config field to BlogLatestEntriesPlugin
 * Fix __str__ plugins method
 * Fix bug when selecting plugins template
 
+******************
 0.6.1 (2015-10-31)
-++++++++++++++++++
+******************
 
 * Improve toolbar: add all languages for each post
 * Improve toolbar: add per-apphook configurable changefreq, priority
 
+******************
 0.6.0 (2015-10-30)
-++++++++++++++++++
+******************
 
 * Add support for django CMS 3.2 Wizard
 * Add support for Apphook Config
@@ -65,8 +123,9 @@ History
 * LatestPostsPlugin tags field has been changed to a plain TaggableManager field.
   A migration is in place to move the data, but backup your data first.
 
+******************
 0.5.0 (2015-08-09)
-++++++++++++++++++
+******************
 
 * Add support for Django 1.8
 * Drop dependency on Django select2
@@ -75,8 +134,9 @@ History
 * Add categories menu
 * Add option to disable the abstract
 
+******************
 0.4.0 (2015-03-22)
-++++++++++++++++++
+******************
 
 * Fix Django 1.7 issues
 * Fix dependencies on python 3 when using wheel packages
@@ -84,14 +144,16 @@ History
 * Fix various templates issues
 * UX fixes in the admin
 
+******************
 0.3.1 (2015-01-07)
-++++++++++++++++++
+******************
 
 * Fix page_name in template
 * Set cascade to set null for post image and thumbnail options
 
+******************
 0.3.0 (2015-01-04)
-++++++++++++++++++
+******************
 
 * Multisite support
 * Configurable default author support
@@ -107,9 +169,9 @@ History
 * Django 1.7 support
 * Python 3.3 and 3.4 support
 
-
+******************
 0.2.0 (2014-09-24)
-++++++++++++++++++
+******************
 
 * **INCOMPATIBLE CHANGE**: view names changed!
 * Based on django parler 1.0
@@ -121,8 +183,8 @@ History
 * Simpler TextField-based content editing for simpler blogs
 * Add support for custom user models
 
-
+******************
 0.1.0 (2014-03-06)
-++++++++++++++++++
+******************
 
 * First experimental release